Extracted expression parser
