+++ tests/handle_parse_error.h
@@ -1,6 +1,7 @@
 #pragma once
 #include "lpg_find_next_token.h"
 #include "lpg_parse_expression.h"
+#include "lpg_expression_parser.h"
 
 typedef struct test_parser_user
 {
